Symmetrized code
One option for making PF symmetric. In this version, the adult individual only reproduces if both single cells reproduce. It is also possible to play with versions where the adult individual reproduces if one of the cells reproduces (this then induces the other cell to reproduce), which increases reproductive rate and allows the two-celled individual to reproduce twice in one time step similarly to the original PF.